Move bughunt runtime setup into module
